backend/requirements.txt
elevenlabs
python-dotenv
langchain_openai
langchain_core
pydantic
langchain_community
pipecat-ai
dotenv
fastapi
pipecat-ai-cli
pipecat-ai[cartesia,daily,deepgram,elevenlabs,openai,runner,silero,webrtc]
torch
uvicorn
twilio
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pytest
pytest-asyncio